Fixed bug loading multiple files

+++ src/cell.h
@@ -46,7 +46,7 @@ extern const struct cell TRUE;
 #define is_compound(e) (is_pair(e) || is_vector(e))
 
 #define mk_cell(t, v) ((struct cell){t, {v}})
-#define mk_num(num) mk_cell(INTEGER, (void *)num)
+#define mk_num(num) mk_cell(INTEGER, (void *)(long)num)
 #define mk_str(str) mk_cell(STRING, strdup(str))
 #define mk_sym(str) mk_cell(SYMBOL, strdup(str))
 #define mk_prim(func) mk_cell(PRIMITIVE, func)
